Fix Android build: two compiler errors in sandbox webview and KERI bridge
sandbox_webview_native.dart:
- InAppWebView and InAppWebViewController are only in the flutter_inappwebview
  meta-package (which we exclude to prevent web plugin crash). Switch to the
  platform_interface factory API: InAppWebViewPlatform.instance
  .createPlatformInAppWebViewWidget(PlatformInAppWebViewWidgetCreationParams)
  .build(context). This compiles on Android/iOS without the meta-package.
- Change controller field type to PlatformInAppWebViewController (from
  platform_interface, available on all platforms).

keri_bridge.rs + keri_bridge.dart:
- Change InteractResult.sequence_number from u64 to i64. FRB v2 maps u64
  to BigInt in Dart but i64 maps to int. Fixes BigInt assignment error.
- Add defensive cast in keri_bridge.dart wrapper as belt-and-suspenders.

Web build verified clean. Both fixes compile without touching runtime behavior.

+++ identity_agent_ui/rust/src/api/keri_bridge.rs
@@ -194,7 +194,7 @@ pub fn verify_signature(data: Vec<u8>, signature: String, public_key: String) ->
 pub struct InteractResult {
     pub aid: String,
     pub said: String,
-    pub sequence_number: u64,
+    pub sequence_number: i64,
     pub kel_entry: String,
     pub raw_bytes_b64: String,
 }
@@ -211,7 +211,7 @@ pub fn interact_aid(name: String, seal_data_json: String) -> Result<InteractResu
         .get_mut(&name)
         .ok_or_else(|| format!("No AID found with name: {}", name))?;
 
-    let sn = instance.kel.len() as u64;
+    let sn = instance.kel.len() as i64;
     let aid = instance.prefix.to_string();
 
     // Parse seal data (pass-through; caller provides validated JSON array).
@@ -223,7 +223,7 @@ pub fn interact_aid(name: String, seal_data_json: String) -> Result<InteractResu
         "t": "ixn",
         "d": "",
         "i": aid,
-        "s": format!("{:x}", sn),
+        "s": format!("{:x}", sn as u64),
         "p": instance.kel.last().cloned().unwrap_or_default(),
         "a": seal_data,
     });
